Refactor -- misc. corrections.
- Added missing operator= or copy constructors, in the cases one exists but not the other.
- Removed some constructors that aren't doing anything special.
- Deleted unused function declarations and variables.
- Corrected some comments.
- Removed an unsigned long since a size_t is already the same size.
- Other small misc. corrections.

+++ Source/source/drawing.cpp
@@ -591,7 +591,7 @@ void gameplay::draw_hud() {
     if(
         hud_items.get_draw_data(HUD_ITEM_PIKMIN_TOTAL_NR, &i_center, &i_size)
     ) {
-        unsigned long total_pikmin = pikmin_list.size();
+        size_t total_pikmin = pikmin_list.size();
         for(size_t o = 0; o < onions.size(); ++o) {
             for(size_t m = 0; m < N_MATURITIES; ++m) {
                 total_pikmin += onions[o]->pikmin_inside[m];
